feat: 예약정보 조회 기능 구현

+++ src/main/java/com/pyeonrimium/queuing/reservation/domains/ReservationRequest.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 @Setter
 public class ReservationRequest {
 
-	private int storeId;
+	private Long storeId;
 	private String storeName;
 	private int partySize;
 	private String request;
